# backend/routes/plans.py
"""Action Plan API Routes — REST endpoints for V2 action plan system."""
from fastapi import APIRouter, Request, HTTPException
from typing import Optional
from pydantic import BaseModel
import logging

from deps import get_current_user, db
from services import (
    action_plan_manager,
    portfolio_intelligence,
    signal_detector,
)

router = APIRouter(prefix="/api")
plan_manager = action_plan_manager.get_plan_manager()
logger = logging.getLogger(__name__)


# ══════════════════════════════════════════════════════════════════════════
# REQUEST/RESPONSE MODELS
# ══════════════════════════════════════════════════════════════════════════

class ActionStatusUpdate(BaseModel):
    """Request body for updating action status."""
    status: str  # "PENDING" | "IN_PROGRESS" | "COMPLETED" | "SKIPPED"
    completion_note: Optional[str] = None


class ActionFeedback(BaseModel):
    """Request body for action feedback."""
    useful: bool  # True = 👍, False = 👎
    comment: Optional[str] = None


# ── v4 — POST /plans/active/actions (cross-domain action create) ──────────
# Per docs/api-changes.md New Endpoint B.10. Used by Tax / Goals / advisor
# Discuss promotion to add actions that the standard plan generator did not
# auto-create. All new v4 fields per Decision 2 (snake_case).

class CreateActionRequest(BaseModel):
    source_domain: str  # concentration|diversification|risk|performance|goals|tax
    type: str           # EXIT|ADD|TRIM|HOLD|REVIEW|SWITCH|MERGE|HARVEST|RAISE_SIP
    verb: Optional[str] = None
    asset_type: str     # STOCK|FUND|DEBT|GOLD|portfolio|mutual_fund|equity
    asset_name: str
    instrument_id: Optional[str] = None
    amount: float = 0
    priority: int = 2
    priority_label: Optional[str] = None
    impact: Optional[dict] = None        # {label, value}
    effort: str = "MEDIUM"               # LOW|MEDIUM|HIGH
    trade_off: str = ""                  # mandatory per PRD §7.4
    expected_impact: Optional[dict] = None
    exclusive: bool = False
    reason_text: str = ""
    reason_codes: list[str] = []
    confidence: str = "MEDIUM"           # HIGH|MEDIUM|LOW


class DiscussActionRequest(BaseModel):
    """Per PRD §10.2 — advisor records a Discuss event without touching status."""
    discussion_note: Optional[str] = None


# ══════════════════════════════════════════════════════════════════════════
# PLAN GENERATION
# ══════════════════════════════════════════════════════════════════════════

@router.post("/plans/generate")
async def generate_plan(request: Request):
    """Generate a new action plan.
    
    Flow:
    1. Fetch portfolio data
    2. Run portfolio intelligence
    3. Generate signals
    4. Calculate scores
    5. Create plan (status="preview")
    6. Return plan for user review
    
    Returns:
        {
            "plan": {...},
            "message": "Plan generated successfully"
        }
    """
    user = await get_current_user(request)
    user_id = user["user_id"]
    
    # Fetch portfolio data
    holdings = await db.holdings.find({"user_id": user_id}, {"_id": 0}).to_list(500)
    
    if not holdings:
        raise HTTPException(status_code=400, detail="No holdings found. Upload your portfolio first.")
    
    # Run portfolio intelligence
    intelligence = await portfolio_intelligence.compute_portfolio_intelligence(user_id)
    
    # Generate plan using ActionPlanManager
    plan = await plan_manager.generate_plan(user_id, intelligence, holdings)
    
    # Save as preview
    await plan_manager.create_plan(plan)
    
    return {
        "plan": plan,
        "message": "Plan generated successfully. Review and save to activate."
    }


# ══════════════════════════════════════════════════════════════════════════
# PLAN CRUD
# ══════════════════════════════════════════════════════════════════════════

@router.post("/plans/{plan_id}/save")
async def save_plan(plan_id: str, request: Request):
    """Save a preview plan as active.
    
    Changes status from "preview" to "active"
    """
    user = await get_current_user(request)
    user_id = user["user_id"]
    
    try:
        plan = await plan_manager.save_plan(plan_id, user_id)
        return {
            "plan": plan,
            "message": "Plan saved successfully"
        }
    except ValueError as e:
        raise HTTPException(status_code=404, detail=str(e))


@router.get("/plans/active")
async def get_active_plan(request: Request, source_domain: Optional[str] = None):
    """Get user's active plan.

    Also runs auto-archive for old completed plans.

    Query params:
        source_domain — v4 filter; one of
            concentration | diversification | risk | performance | goals | tax.
            When set, returns only actions tagged with that source_domain
            (push-down filter for the per-dashboard recommendation matrix).

    Returns:
        {
            "plan": {...} | null,
            "has_plan": true | false
        }
    """
    user = await get_current_user(request)
    user_id = user["user_id"]

    # Auto-archive old completed plans (runs in background)
    try:
        await plan_manager.auto_archive_old_completed_plans(user_id)
    except Exception as e:
        # Don't fail the request if auto-archive fails
        logger.warning("Auto-archive failed: %s", e)

    plan = await plan_manager.get_active_plan(user_id, source_domain=source_domain)

    return {
        "plan": plan,
        "has_plan": plan is not None
    }


@router.delete("/plans/active")
async def clear_active_plan(request: Request):
    """Archive the current active plan.

    Used by the trash button in the Plan Board UI when the user wants
    to discard the current recommendations and start over (e.g. after
    significant portfolio changes, or when the suggestions feel stale).
    A completely fresh plan can then be generated via POST /plans/generate.

    Returns the archived plan's id + status so the UI can confirm.
    Idempotent — calling on a user with no active plan returns ok=False.
    """
    user = await get_current_user(request)
    user_id = user["user_id"]

    plan = await plan_manager.get_active_plan(user_id)
    if not plan:
        return {"ok": False, "reason": "no_active_plan",
                "message": "No active plan to clear."}
    archived = await plan_manager.archive_plan(
        plan["plan_id"], user_id, reason="cleared_by_user",
    )
    if not archived:
        return {"ok": False, "reason": "archive_failed",
                "plan_id": plan["plan_id"]}
    return {
        "ok": True,
        "plan_id": plan["plan_id"],
        "version": plan.get("version"),
        "message": "Active plan cleared. Generate a new plan to get fresh recommendations.",
    }


@router.get("/plans/active/health-projection")
async def get_active_plan_health_projection(request: Request):
    """Return current + projected Portfolio Health if pending actions are
    implemented.

    Response shape: see `services.portfolio_health_projection.project_health`.
    """
    user = await get_current_user(request)
    try:
        from services import portfolio_health_projection as _phproj
        result = await _phproj.project_health(user["user_id"])
    except Exception as e:  # noqa: BLE001
        logger.warning("health projection failed: %s", e)
        return {"available": False, "reason": str(e)}
    if result is None:
        return {"available": False, "reason": "No active plan or no holdings."}
    return {"available": True, **result}


@router.get("/plans/history")
async def get_plan_history(
    request: Request,
    limit: int = 10,
    skip: int = 0,
):
    """Get user's plan history."""
    user = await get_current_user(request)
    user_id = user["user_id"]
    
    plans = await plan_manager.get_plan_history(user_id, limit, skip)
    
    return {
        "plans": plans,
        "count": len(plans)
    }


@router.get("/plans/{plan_id}")
async def get_plan(plan_id: str, request: Request):
    """Get specific plan by ID."""
    user = await get_current_user(request)
    user_id = user["user_id"]
    
    plan = await plan_manager.get_plan(plan_id, user_id)
    
    if not plan:
        raise HTTPException(status_code=404, detail="Plan not found")
    
    return {"plan": plan}


# ══════════════════════════════════════════════════════════════════════════
# ACTION STATUS MANAGEMENT
# ══════════════════════════════════════════════════════════════════════════

@router.patch("/plans/{plan_id}/actions/{action_id}")
async def update_action_status(
    plan_id: str,
    action_id: str,
    body: ActionStatusUpdate,
    request: Request,
):
    """Update action status.
    
    Body:
        {
            "status": "PENDING" | "IN_PROGRESS" | "COMPLETED" | "SKIPPED",
            "completion_note": "Redeemed via Groww app" (optional)
        }
    
    Returns:
        Updated plan with new progress
    """
    user = await get_current_user(request)
    user_id = user["user_id"]
    
    try:
        updated_plan = await plan_manager.update_action_status(
            plan_id=plan_id,
            action_id=action_id,
            user_id=user_id,
            new_status=body.status,
            completion_note=body.completion_note,
        )
        
        return {
            "plan": updated_plan,
            "message": f"Action {action_id} marked as {body.status}"
        }
    except ValueError as e:
        raise HTTPException(status_code=404, detail=str(e))


# ══════════════════════════════════════════════════════════════════════════
# v4 — CROSS-DOMAIN ACTION CREATE (B.10) + ADVISOR DISCUSS (B.11)
# ══════════════════════════════════════════════════════════════════════════

@router.post("/plans/active/actions", status_code=201)
async def create_active_plan_action(body: CreateActionRequest, request: Request):
    """Create a new action on the active plan.

    Per docs/api-changes.md B.10 — used when the source is a widget (Tax)
    or a domain that doesn't feed action_plan_manager auto-generation
    (Goals), or for advisor "Discuss" promotion.

    Idempotency: pass `Idempotency-Key: <uuid>` header to make repeated
    POSTs from the same client idempotent.
    """
    user = await get_current_user(request)
    user_id = user["user_id"]
    idempotency_key = request.headers.get("Idempotency-Key")

    try:
        new_action = await plan_manager.add_action_to_active_plan(
            user_id=user_id,
            action_input=body.model_dump(),
            idempotency_key=idempotency_key,
        )
        return new_action
    except ValueError as e:
        raise HTTPException(status_code=404, detail=str(e))


@router.patch("/plans/{plan_id}/actions/{action_id}/discuss")
async def discuss_action(
    plan_id: str,
    action_id: str,
    body: DiscussActionRequest,
    request: Request,
):
    """Advisor-side Discuss-only write.

    Per PRD §10.2 + docs/api-changes.md B.11. Records that the advisor
    discussed this action with the client; does NOT change action.status.
    The client retains sole control over Accept/Skip.
    """
    user = await get_current_user(request)
    user_id = user["user_id"]

    try:
        updated_action = await plan_manager.record_action_discussion(
            plan_id=plan_id,
            action_id=action_id,
            advisor_user_id=user_id,
            note=body.discussion_note,
        )
        return updated_action
    except ValueError as e:
        raise HTTPException(status_code=404, detail=str(e))


@router.patch("/plans/{plan_id}/actions/{action_id}/feedback")
async def submit_action_feedback(
    plan_id: str,
    action_id: str,
    body: ActionFeedback,
    request: Request,
):
    """Submit feedback for an action.
    
    Body:
        {
            "useful": true | false,
            "comment": "Optional feedback text"
        }
    
    Returns:
        Updated plan with feedback recorded
    """
    user = await get_current_user(request)
    user_id = user["user_id"]
    
    try:
        updated_plan = await plan_manager.update_action_feedback(
            plan_id=plan_id,
            action_id=action_id,
            user_id=user_id,
            useful=body.useful,
            comment=body.comment,
        )
        
        return {
            "plan": updated_plan,
            "message": "Feedback recorded successfully"
        }
    except ValueError as e:
        raise HTTPException(status_code=404, detail=str(e))


# ══════════════════════════════════════════════════════════════════════════
# PLAN REFRESH
# ══════════════════════════════════════════════════════════════════════════

@router.post("/plans/refresh")
async def refresh_plan(request: Request):
    """Refresh active plan (generate new version).
    
    Flow:
    1. Archive current active plan
    2. Generate new plan with incremented version
    3. Save as active
    
    Use when:
    - Portfolio has changed significantly
    - User wants updated recommendations
    """
    user = await get_current_user(request)
    user_id = user["user_id"]
    
    # Fetch latest portfolio data
    holdings = await db.holdings.find({"user_id": user_id}, {"_id": 0}).to_list(500)
    
    if not holdings:
        raise HTTPException(status_code=400, detail="No holdings found")
    
    # Run portfolio intelligence (fetched but not returned - used by plan_manager internally)
    await portfolio_intelligence.compute_portfolio_intelligence(user_id)
    
    # Refresh plan
    new_plan = await plan_manager.refresh_plan(user_id)
    
    return {
        "plan": new_plan,
        "message": f"Plan refreshed to version {new_plan['version']}"
    }


# ══════════════════════════════════════════════════════════════════════════
# SIGNALS (for UI)
# ══════════════════════════════════════════════════════════════════════════

@router.get("/signals/generate")
async def generate_signals(request: Request):
    """Generate portfolio signals without creating a full plan.
    
    Useful for showing signals in collapsible widget.
    
    Returns:
        {
            "signals": [...],
            "summary": {...}
        }
    """
    user = await get_current_user(request)
    user_id = user["user_id"]
    
    # Fetch portfolio data
    holdings = await db.holdings.find({"user_id": user_id}, {"_id": 0}).to_list(500)
    
    if not holdings:
        return {
            "signals": [],
            "summary": {
                "total_signals": 0,
                "high_severity": 0,
                "medium_severity": 0,
                "low_severity": 0,
            }
        }
    
    # Run portfolio intelligence
    intelligence = await portfolio_intelligence.compute_portfolio_intelligence(user_id)
    
    # Generate signals
    portfolio_data = {
        "portfolio_intelligence": intelligence,
        "holdings": holdings,
        "total_value": sum(h["quantity"] * h["current_price"] for h in holdings),
    }
    
    signals = signal_detector.generate_signals(portfolio_data)
    summary = signal_detector.get_signal_summary(signals)
    
    return {
        "signals": signals,
        "summary": summary
    }


# ══════════════════════════════════════════════════════════════════════════
# PLAN SIMULATION (Future)
# ══════════════════════════════════════════════════════════════════════════

@router.post("/plans/{plan_id}/simulate")
async def simulate_plan(plan_id: str, request: Request):
    """Simulate plan impact (before/after comparison).
    
    For MVP: Placeholder
    In production: Show portfolio metrics before/after executing plan
    """
    user = await get_current_user(request)
    user_id = user["user_id"]
    
    plan = await plan_manager.get_plan(plan_id, user_id)
    
    if not plan:
        raise HTTPException(status_code=404, detail="Plan not found")
    
    # For MVP: Return basic simulation
    return {
        "simulation": {
            "before": {
                "total_value": plan["metadata"]["portfolio_value_at_creation"],
                "overlap_score": 30.4,  # Placeholder
                "concentration_risk": "HIGH",
            },
            "after": {
                "total_value": plan["metadata"]["portfolio_value_at_creation"] - plan["freed_capital"] + plan["post_tax_proceeds"],
                "overlap_score": 45.0,  # Improved
                "concentration_risk": "MEDIUM",
            },
            "improvements": [
                "Overlap reduced by 15 points",
                "Concentration risk lowered",
                "Tax-efficient exits",
            ]
        },
        "message": "Simulation complete"
    }



@router.post("/plans/refresh-fundamentals")
async def refresh_fundamentals(request: Request):
    """Refresh fundamental data cache for all user's stock holdings.
    
    This endpoint:
    1. Fetches latest fundamental data from Groww for all stocks
    2. Caches data in Redis (24h expiry)
    3. Returns count of successfully updated stocks
    
    Should be called:
    - Once daily (can be automated)
    - Before generating a new plan
    - On user request
    """
    from services.groww_fundamentals import refresh_cache_for_user
    
    user = await get_current_user(request)
    user_id = user["user_id"]
    
    try:
        fundamentals_map = await refresh_cache_for_user(user_id)
        
        return {
            "success": True,
            "stocks_updated": len(fundamentals_map),
            "updated_stocks": list(fundamentals_map.keys()),
            "message": f"Successfully refreshed fundamental data for {len(fundamentals_map)} stocks"
        }
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))