# docs/backend/nivesh-app-openapi.yaml
openapi: 3.1.0
info:
  title: Nivesh App API
  version: 2.0.0
  description: |
    REST API for the Nivesh.ai application — portfolio management, goal planning,
    AI copilot, broker integration, MFD workspace, and admin controls.

    ## Authentication

    All endpoints (except `/api/auth/*`, `/api/public/*`, and `/api/cas-parser-provider/active`)
    require a session cookie set by `POST /api/auth/google`.

    Admin endpoints additionally require the requesting user to have `is_admin = true`.

    ## Base URL

    - **Production**: `https://niveshcopilot.com`
    - **Staging**: `https://staging.niveshcopilot.com`

  contact:
    name: Nivesh
    url: https://nivesh.com
servers:
  - url: https://niveshcopilot.com
    description: Production
  - url: https://staging.niveshcopilot.com
    description: Staging
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: Session cookie set by `POST /api/auth/google`.
...
security:
  - SessionCookie: []
paths:
...
  /api/plans/history:
    get:
      tags: [plans]
      operationId: plans_history
      summary: Get plan history
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: skip
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Plan history
...
  /api/plans/{plan_id}:
    get:
      tags: [plans]
      operationId: plans_get
      summary: Get a specific plan by ID
      parameters:
        - name: plan_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Plan document
  /api/plans/{plan_id}/save:
    post:
      tags: [plans]
      operationId: plans_save
      summary: Save a preview plan as the active plan
      parameters:
        - name: plan_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Saved plan
  /api/plans/{plan_id}/simulate:
    post:
      tags: [plans]
      operationId: plans_simulate
      summary: Simulate before/after impact of a plan
      parameters:
        - name: plan_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Simulation result
  /api/plans/{plan_id}/actions/{action_id}:
    patch:
      tags: [plans]
      operationId: plans_update_action
      summary: Update action status (pending/done/skipped)
      parameters:
        - name: plan_id
          in: path
          required: true
          schema:
            type: string
        - name: action_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [pending, done, skipped]
                completion_note:
                  type: string
                  nullable: true
      responses:
        "200":
          description: Updated plan
  /api/plans/{plan_id}/actions/{action_id}/feedback:
    patch:
      tags: [plans]
      operationId: plans_action_feedback
      summary: Submit feedback on an action recommendation
      parameters:
        - name: plan_id
          in: path
          required: true
          schema:
            type: string
        - name: action_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [useful]
              properties:
                useful:
                  type: boolean
                comment:
                  type: string
                  nullable: true
      responses:
        "200":
          description: Updated plan
...
  /api/copilot/agents/route:
    post:
      tags: [copilot_agents]
      operationId: copilot_route_intent
      summary: Deterministic intent router — classify message to agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message:
                  type: string
      responses:
        "200":
          description: Routed agent ID and spec
...